Refactor: cts: Move patterns.py into the _cts module.
This just moves the file and fixes up imports of it elsewhere.  No
further code modifications have taken place.
