+++ python/pacemaker/_cts/Makefile.am
@@ -14,5 +14,6 @@ pkgpythondir = $(pythondir)/$(PACKAGE)/_cts
 pkgpython_PYTHON = __init__.py \
 				   corosync.py \
 				   errors.py \
+				   patterns.py \
 				   process.py \
 				   test.py
