+++ python/pacemaker/_cts/patterns.py
